Refactor import paths and add type annotations for ThemeNode in TemasideHubPage and TemasideIndexPage
