+++ src/pages/temasider/TemasideIndexPage.tsx
@@ -1,6 +1,7 @@
 import { Link } from "react-router-dom";
 import { Heading } from "@digdir/designsystemet-react";
-import { getTopLevelTemasider } from "../../lib/temasider/temasiderTree";
+import { getTopLevelTemasider } from "../../lib/temaside/temasiderTree";
+import type { ThemeNode } from "../../lib/temaside/temasiderTree";
 
 export function TemasideIndexPage() {
   const top = getTopLevelTemasider();
@@ -10,7 +11,7 @@ export function TemasideIndexPage() {
       <Heading level={1} data-size="lg">Temasider</Heading>
 
       <div className="mt-6 grid gap-3 sm:grid-cols-2 lg:grid-cols-3">
-        {top.map((t) => (
+        {top.map((t: ThemeNode) => (
           <Link
             key={t.path}
             to={`/temasider${t.path}`}
